Fixes tests and examples to run on scala 2.12 and 2.13
