+++ build.sbt
@@ -53,7 +53,7 @@ def scalaVersionSpecificFolders(srcName: String, srcBaseDir: java.io.File, scala
 }
 
 lazy val buildSettings = Seq(
-  scalaVersion       := "2.13.0",
+  scalaVersion       := "2.12.8",
   crossScalaVersions := Seq("2.12.8", "2.13.0"),
   scalatestVersion   := "3.1.0-SNAP13",
   scalacOptions     ++= Seq(
