+++ example/src/test/scala/monocle/function/FilterIndexExample.scala
@@ -9,7 +9,6 @@ class FilterIndexExample extends MonocleSuite {
   test("filterIndexes creates Traversal from a SortedMap, IMap to all values where the index matches the predicate") {
 
     (SortedMap("One" -> 1, "Two" -> 2) applyTraversal filterIndex{k: String => k.toLowerCase.contains("o")} getAll)  shouldEqual List(1, 2)
-
     (SortedMap("One" -> 1, "Two" -> 2) applyTraversal filterIndex{k: String => k.startsWith("T")} set 3) shouldEqual SortedMap("One" -> 1, "Two" -> 3)
 
   }
@@ -20,8 +19,6 @@ class FilterIndexExample extends MonocleSuite {
 
     (List(1,3,5,7)   applyTraversal filterIndex{ i: Int => i >= 2 } modify(_ + 2)) shouldEqual List(1,3,7,9)
     (Vector(1,3,5,7) applyTraversal filterIndex{ i: Int => i >= 2 } modify(_ + 2)) shouldEqual Vector(1,3,7,9)
-    (Stream(1,3,5,7) applyTraversal filterIndex{ i: Int => i >= 2 } modify(_ + 2)) shouldEqual Stream(1,3,7,9)
-
   }
 
 }
