+++ test/shared/src/test/scala-2.12-/monocle/std/StreamSpec.scala
