+++ test/shared/src/test/scala-2.13+/monocle/ScalaVersionSpecificTestInstances.scala
@@ -1,0 +1,3 @@
+package monocle
+
+private [monocle] trait ScalaVersionSpecificTestInstances
