+++ test/shared/src/test/scala/monocle/IsoSpec.scala
@@ -123,20 +123,20 @@ class IsoSpec extends MonocleSuite {
     iso.modify(_ + 1)(IntWrapper(0)) shouldEqual IntWrapper(1)
   }
 
-  test("GenIso nullary equality") {
-    GenIso.unit[Nullary] shouldEqual _nullary
-  }
-
-  test("GenIso unary equality") {
-    GenIso[Unary, Int] shouldEqual _unary
-  }
-
-  test("GenIso binary equality") {
-    GenIso.fields[Binary] shouldEqual _binary
-  }
-
-  test("GenIso quintary equality") {
-    GenIso.fields[Quintary] shouldEqual _quintary
-  }
+  // test("GenIso nullary equality") {
+  //   GenIso.unit[Nullary] shouldEqual _nullary
+  // }
+  //
+  // test("GenIso unary equality") {
+  //   GenIso[Unary, Int] shouldEqual _unary
+  // }
+  //
+  // test("GenIso binary equality") {
+  //   GenIso.fields[Binary] shouldEqual _binary
+  // }
+  //
+  // test("GenIso quintary equality") {
+  //   GenIso.fields[Quintary] shouldEqual _quintary
+  // }
 
 }
